requirements.txt
flask
twilio
google-generativeai
requests

